Fixed fall damage
* Fixes #3216

+++ src/Entities/Player.cpp
@@ -1439,7 +1439,7 @@ void cPlayer::TeleportToCoords(double a_PosX, double a_PosY, double a_PosZ)
 	{
 		SetPosition(a_PosX, a_PosY, a_PosZ);
 		FreezeInternal(GetPosition(), false);
-		m_LastGroundHeight = static_cast<float>(a_PosY);
+		m_LastGroundHeight = a_PosY;
 		m_bIsTeleporting = true;
 
 		m_World->BroadcastTeleportEntity(*this, GetClientHandle());
